Extract UserPicker component from InviteMemberDialog
